extracted team and settings subcommands, moved and renamed jam admin commands
