+++ bot/src/main/java/de/chojo/gamejam/commands/SettingsSubcommand.java
@@ -1,0 +1,10 @@
+package de.chojo.gamejam.commands;
+
+import de.chojo.gamejam.data.wrapper.jam.JamSettings;
+import de.chojo.jdautil.wrapper.SlashCommandContext;
+import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
+
+public interface SettingsSubcommand {
+
+    void execute(SlashCommandInteractionEvent event, SlashCommandContext context, JamSettings settings);
+}
