+++ bot/src/main/java/de/chojo/gamejam/commands/TeamSubCommand.java
@@ -1,0 +1,9 @@
+package de.chojo.gamejam.commands;
+
+import de.chojo.gamejam.data.wrapper.jam.Jam;
+import de.chojo.jdautil.wrapper.SlashCommandContext;
+import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
+
+public interface TeamSubCommand {
+    void execute(SlashCommandInteractionEvent event, SlashCommandContext context, Jam jam);
+}
